requirements.txt
#requirements
#Core UI and app
streamlit==1.32.2
pandas==2.3.0
matplotlib==3.10.3
openpyxl==3.1.2
fpdf==1.7.2
jsonlines==3.1.0

#Semantic Search
sentence-transformers==2.7.0
faiss-cpu==1.11.0
scikit-learn==1.5.0
numpy==1.26.4
torch==2.7.1
transformers

#LangChain and Ollama (LLM Support)
langchain==0.3.25
langchain_core==0.3.65
langchain_community==0.0.32
langchain_ollama==0.3.3
llama-index==0.10.38
llama-index-vector-stores-faiss==0.1.2
llama-index-embeddings-huggingface==0.2.0

